taskproject/requirements.txt
amqp==5.1.1
appier==1.30.1
asgiref==3.6.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
cachetools==3.1.1
celery==5.2.3
redis==4.1.4
django-redis==5.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.2
django-filter==23.1
djangorestframework==3.14.0

httplib2==0.22.0
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
prompt-toolkit==3.0.38
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
